Move argContext enum to separate file
